requirements.txt
flask==3.0.3
pyyaml==6.0
requests==2.30.0
urllib3==1.26.17
pillow==10.0.0
cryptography==41.0.0
